--- src/components/header/Header.module.css
@@ -1,191 +1,0 @@
-.sticky_header {
-    display: flex;
-    align-items: flex-start;
-    justify-content: space-between;
-    position: fixed;
-    top: 0;
-    right: 0;
-    left:0;
-    z-index: 3;
-}
-
-.header_sections {
-    backdrop-filter: blur(10px);
-    background-color: rgba(19, 19, 23, 0.7);
-    color: var(--text-color-primary);
-    position: fixed;
-    display: flex;
-    justify-content: center;
-    padding-top: 80px;
-    padding-bottom: 80px;
-    width: 100%;
-    z-index: 2;
-    border-bottom: 2px solid var(--color-beige-secondary);
-}
-
-.header_title {
-    align-self: center;
-    font-family: var(--font-family-secondary);
-    font-size: 1.5rem;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
-}
-
-.container_left {
-    display: flex;
-    align-items: center;
-    column-gap: var(--spacing-XXXS);
-    padding-top: var(--spacing-XS);
-    padding-bottom: var(--spacing-XS);
-    padding-right: var(--spacing-XS);
-    padding-left: var(--spacing-S);
-}
-
-.container_right {
-    display: flex;
-    align-items: center;
-    padding-top: var(--spacing-L);
-    padding-right: var(--spacing-S);
-}
-
-.header_button {
-    -webkit-tap-highlight-color: transparent;
-    font-family: var(--font-family-secondary);
-    color: var(--text-color-primary);
-    background-color: transparent;
-    border: none;
-    cursor: pointer;
-    .button_text {
-        display: none;
-    }
-    .hamburger_icon {
-        border: 2px solid var(--background-color-primary);
-        width: 40px;
-        height: 40px;
-        border-radius: 50%;
-        padding: 5px;
-    }
-
-    .animate {
-        animation: scaleAndFade 0.5s ease;
-    }
-}
-@keyframes scaleAndFade {
-    0%, 100% {
-        opacity: 1;
-    }
-    50% {
-        opacity: 0.5;
-    }
-}
-
-/* Styles for large screens */
-@media screen and (min-width: 768px) {
-    .header_button  {
-        height: 40px;
-        border: 2px solid var(--background-color-primary);
-        border-radius: 8px;
-        padding-left: var(--spacing-XS);
-        padding-right: var(--spacing-XS);
-        .hamburger_icon {
-            display: none;
-        }
-        .button_text {
-            display: block;
-        }
-
-        &:hover {
-            border-color: var(--color-cyan-secondary);
-        }
-    }
-}
-
-/* Disable hover from touch devices */
-@media (hover: none) and (pointer: coarse) {
-    .header_button {
-        &:hover {
-            background-color: initial;
-        }
-    }
-}
-
-.open {
-    transform: translateY(0);
-    transition: transform 0.3s;
-}
-  
-.closed {
-    transform: translateY(-68%);
-    transition: transform 0.3s;
-}
-
-.custom_colors {
-    display: flex;
-    flex-direction: column;
-    list-style: none;
-    margin-right: 3%;
-    margin-top: 0px;
-    margin-bottom: 0px;
-    counter-reset: list-index;
-}
-
-.custom_colors li {
-    counter-increment: list-index;
-}
-  
-.custom_colors li::before {
-    content: counter(list-index) ". ";
-    font-family: var(--font-family-primary);
-    color: var(--text-color-orangeRed-secondary);
-}
-  
-.custom_colors li:nth-child(1)::before {
-    color: var(--text-color-lightCyan-secondary);
-}
-  
-.custom_colors li:nth-child(2)::before {
-    color: var(--text-color-lightGreen-secondary);
-}
-  
-.custom_colors li:nth-child(3)::before {
-    color: var(--text-color-orange-secondary);
-}
-  
-.li_item {
-    transition: color 0.3s ease;
-    position: relative;
-    padding-bottom: var(--spacing-XXXS);
-    cursor: pointer;
-    font-family: var(--font-family-secondary);
-}
-  
-.li_item:hover {
-    color: var(--text-color-cyan-secondary);
-}
-  
-.li_item::after {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 0;
-    height: 2px;
-    background-color: var(--color-darkCyan-secondary);
-    transition: width 0.3s ease;
-}
-  
-.li_item:hover::after {
-    width: 100%;
-}
-
-/* Override for touch devices */
-@media (hover: none) and (pointer: coarse) {
-    .li_item {
-        -webkit-tap-highlight-color: transparent;
-    }
-    .li_item:hover::after {
-      background-color: initial;
-      width: 0;
-    }
-}
