+++ src/components/header/sections/HeaderSections.tsx
@@ -2,7 +2,7 @@ import { HeaderSectionProps } from "src/components/header/types";
 import { useSectionTitleIdContext } from "src/components/header/SectionIdProvider";
 import { Section } from "src/components/app/types";
 import { useTranslation } from "react-i18next";
-import * as styles from "src/components/header/Header.module.css";
+import * as styles from "src/components/header/sections/HeaderSections.module.css";
 
 export const HeaderSections = ({
   onClick,
