+++ src/components/header/toolbar/HeaderToolbar.tsx
@@ -1,13 +1,13 @@
-import * as styles from "src/components/header/Header.module.css";
+import * as styles from "src/components/header/toolbar/HeaderToolbar.module.css";
 import { RectangleProgress } from "src/components/rectangleProgress/RectangleProgress";
-import { HeaderToggleProps } from "src/components/header/types";
+import { HeaderToolbarProps } from "src/components/header/types";
 import { LanguageSelector } from "src/components/languageSelector/LanguageSelector";
-import { Title } from "src/components/header/Title";
-import { MenuButton } from "src/components/header/MenuButton";
+import { Title } from "src/components/header/toolbar/title/Title";
+import { MenuButton } from "src/components/header/toolbar/menuButton/MenuButton";
 
 
 
-export const HeaderToggle = (props: HeaderToggleProps): React.JSX.Element => (
+export const HeaderToolbar = (props: HeaderToolbarProps): React.JSX.Element => (
   <div className={styles.sticky_header}>
     <div className={styles.container_left}>
       <RectangleProgress text={"S.H"} size={60} />
