+++ src/components/header/toolbar/menuButton/MenuButton.module.css
@@ -1,0 +1,61 @@
+.header_button {
+  -webkit-tap-highlight-color: transparent;
+  font-family: var(--font-family-secondary);
+  color: var(--text-color-primary);
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  .button_text {
+      display: none;
+  }
+  .hamburger_icon {
+      border: 2px solid var(--background-color-primary);
+      width: 40px;
+      height: 40px;
+      border-radius: 50%;
+      padding: 5px;
+  }
+
+  .animate {
+      animation: scaleAndFade 0.5s ease;
+  }
+}
+
+@keyframes scaleAndFade {
+  0%, 100% {
+      opacity: 1;
+  }
+  50% {
+      opacity: 0.5;
+  }
+}
+
+/* Styles for large screens */
+@media screen and (min-width: 768px) {
+  .header_button  {
+      height: 40px;
+      border: 2px solid var(--background-color-primary);
+      border-radius: 8px;
+      padding-left: var(--spacing-XS);
+      padding-right: var(--spacing-XS);
+      .hamburger_icon {
+          display: none;
+      }
+      .button_text {
+          display: block;
+      }
+
+      &:hover {
+          border-color: var(--color-cyan-secondary);
+      }
+  }
+}
+
+/* Disable hover from touch devices */
+@media (hover: none) and (pointer: coarse) {
+  .header_button {
+      &:hover {
+          background-color: initial;
+      }
+  }
+}
